actions: Split out zerver.actions.streams.
