+++ zerver/tests/test_audit_log.py
@@ -8,6 +8,13 @@ from django.utils.timezone import now as timezone_now
 from analytics.models import StreamCount
 from zerver.actions.realm_icon import do_change_icon_source
 from zerver.actions.realm_playgrounds import do_add_realm_playground, do_remove_realm_playground
+from zerver.actions.streams import (
+    bulk_add_subscriptions,
+    bulk_remove_subscriptions,
+    do_change_subscription_property,
+    do_deactivate_stream,
+    do_rename_stream,
+)
 from zerver.actions.user_settings import (
     do_change_avatar_fields,
     do_change_password,
@@ -18,8 +25,6 @@ from zerver.actions.user_settings import (
 )
 from zerver.actions.users import do_change_user_role, do_deactivate_user
 from zerver.lib.actions import (
-    bulk_add_subscriptions,
-    bulk_remove_subscriptions,
     do_activate_mirror_dummy_user,
     do_add_realm_domain,
     do_change_bot_owner,
@@ -27,14 +32,11 @@ from zerver.lib.actions import (
     do_change_default_events_register_stream,
     do_change_default_sending_stream,
     do_change_realm_domain,
-    do_change_subscription_property,
     do_create_user,
     do_deactivate_realm,
-    do_deactivate_stream,
     do_reactivate_realm,
     do_reactivate_user,
     do_remove_realm_domain,
-    do_rename_stream,
     do_set_realm_authentication_methods,
     do_set_realm_message_editing,
     do_set_realm_notifications_stream,
