+++ zerver/tests/test_message_topics.py
@@ -1,6 +1,6 @@
 from django.utils.timezone import now as timezone_now
 
-from zerver.lib.actions import do_change_stream_permission
+from zerver.actions.streams import do_change_stream_permission
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.models import Message, UserMessage, get_client, get_realm, get_stream
