+++ zerver/tests/test_realm.py
@@ -9,6 +9,7 @@ from django.conf import settings
 from django.utils.timezone import now as timezone_now
 
 from confirmation.models import Confirmation, create_confirmation_link
+from zerver.actions.streams import do_deactivate_stream
 from zerver.lib.actions import (
     do_add_deactivated_redirect,
     do_change_realm_org_type,
@@ -16,7 +17,6 @@ from zerver.lib.actions import (
     do_change_realm_subdomain,
     do_create_realm,
     do_deactivate_realm,
-    do_deactivate_stream,
     do_scrub_realm,
     do_send_realm_reactivation_email,
     do_set_realm_property,
